Added new fictures
